Add Global Style into own file and imported into app.js so it can be used as a component
